greenscreen now does masking
i am having problems connecting the player mask to the color video frame
from the kinect. this.colorPixels should be recieving the pixels from
the video frame (which the debugger is telling me it is), but I am
losing the data once i call a SetData() function to a Texture2D. Halp!