chatted with noah version of code
we have sillouette working, but are missing RGB image from the
colorImageFrame in this.colorPixels. We need to put RGB data from
colorPixels into the silouetted image (i.e. we need a cutout of people).

currently, there is a method in Draw() to force drawing of the RGB image
frame into the top left of the application if this.ColorPixels is mapped
into finalImage.